# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Plataforma de Atendimento Omnichannel (SaaS Multi-Tenant) API",
    "description": "Especificação completa da API Go (Fiber + PostgreSQL RLS + Redis) estilo Chatwoot + Kommo + Flow Engine",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\""
    }
  },
  "paths": {
...
    "/contacts": {
      "get": {
        "summary": "Lista contatos com busca e paginação",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Lista de contatos" } }
      },
      "post": {
        "summary": "Cadastra novo contato",
        "security": [{ "Bearer": [] }],
        "responses": { "201": { "description": "Contato cadastrado" } }
      }
    },
    "/contacts/merge": {
      "post": {
        "summary": "Mescla dois contatos duplicados em um único registro",
        "security": [{ "Bearer": [] }],
        "responses": { "200": { "description": "Contatos mesclados" } }
      }
    },
...
  }
}